backend/orderflow-engine-service/requirements.txt
fastapi
uvicorn[standard]
websockets
websocket-client
psycopg2-binary
pandas
python-binance
asyncpg